Ability -> master

+++ Extension/Component/Imapct/Component_Impact.ImpactData.cs
@@ -1,5 +1,4 @@
 using Cfg.Enum;
-using System;
 
 namespace Aquila.Fight.Impact
 {
@@ -58,7 +57,7 @@ namespace Aquila.Fight.Impact
             /// <summary>
             /// 叠加层数
             /// </summary>
-            public int _stackCount;
+            public ushort _stackCount;
 
             /// <summary>
             /// 叠加层数上限
